Moving deeplink to app component && checking for facebook weirdness
