+++ app/components/Auth/Auth.js
@@ -4,108 +4,10 @@
 // BackeyM - "pee pee poo poo" (02/18/20)
 // dra031cko - "android > ios" (02/19/20)
 // sillyonly - "SOOOOOO what happens when silly have 1800?!" (02/19/20)
-import { firebase } from '@react-native-firebase/auth'
-
-// Styles
+import React, { memo } from 'react'
+import { StyleSheet, Text, View } from 'react-native'
 import * as StyleConstants from '@StyleConstants'
-import { SetInitialUserData } from 'Gruvee/redux/actions/user/UserActions'
-import React, { useEffect } from 'react'
-import { Linking, StyleSheet, Text, View, Platform } from 'react-native'
-import AsyncStorage from '@react-native-community/async-storage'
-
-// Redux
-import { connect } from 'react-redux'
-import { HandleSpotifyDeepLink } from './components/Buttons/actions/SpotifyActions'
 import Buttons from './components/Buttons'
-
-const DEEP_LINK_IN_PROGRESS_FLAG = '@Deep_Link_In_Progress'
-
-// no_neon_one - "Btw I Use Arch!" (02/17/20)
-const Auth = ({ setInitialUserData }) => {
-    useEffect(() => {
-        // TODO: These are used now while we do not have a signout method
-        // firebase.auth().signOut()
-        // AsyncStorage.setItem('@Deep_Link_In_Progress', 'false')
-        // Deep Link Handler
-        if (Platform.OS === 'android') {
-            // Android instaniates multiple activites with deep links
-            // To combat insane calls to out handler, set a flag here to stop it if it's already working
-            AsyncStorage.getItem(DEEP_LINK_IN_PROGRESS_FLAG).then(value => {
-                console.log('GOT DEEPLINKFLAG: ', value)
-                if (value === null || value === 'false') {
-                    // First time running this
-                    Linking.getInitialURL().then(url => {
-                        if (url !== null) {
-                            handleOpenUrl({ url })
-                        }
-                    })
-                }
-            })
-        } else {
-            Linking.addEventListener('url', handleOpenUrl)
-        }
-
-        // Apple Authentication Handler
-        /* if (appleAuth.isSupported) {
-            // If creds are revoked we probably want to some new stuff here
-            return appleAuth.onCredentialRevoked(async () => {
-                console.warn(
-                    'If this function executes, User Credentials have been Revoked'
-                )
-            })
-        } */
-
-        return () => {
-            // Do cleanup here
-            Linking.removeEventListener('url')
-        }
-    }, [])
-
-    // Helpers
-    const handleOpenUrl = async event => {
-        try {
-            let newUserObj = {}
-            AsyncStorage.setItem('@Deep_Link_In_Progress', 'true')
-
-            // Check to see what platform this is coming from
-            if (event.url.includes('spotify_auth')) {
-                // Gets API token object
-                // HumansNotFish - "Team Yaya. Gotta have faith nerds."(02/21/20)
-                newUserObj = await HandleSpotifyDeepLink(event)
-            }
-
-            // After auth, we should always set initial user data and sign via firebase
-            setInitialUserData(newUserObj.user)
-
-            AsyncStorage.setItem('@Deep_Link_In_Progress', 'false')
-        } catch (error) {
-            // TODO: Handle Error
-            console.warn(error)
-
-            AsyncStorage.setItem('@Deep_Link_In_Progress', 'false')
-        }
-    }
-
-    return (
-        <View style={styles.Container}>
-            <Text style={styles.SectionTitle}>Welcome to Grüvee Beta!</Text>
-            <View style={styles.TextContainer}>
-                <Text style={styles.SectionDetail}>
-                    Thanks for taking part in Grüvee Beta for iOS and Android!
-                </Text>
-                <Text style={styles.SectionDetail}>
-                    It's pretty simple, to get started all you need to do is press the "Login With
-                    Spotify" button.
-                </Text>
-                <Text style={styles.SectionDetail}>
-                    This will give you all the fanciness Grüvee has to offer. Please reach out on
-                    the Discord for any feedback and or questions!
-                </Text>
-            </View>
-            <View style={styles.ButtonContainer}>{Buttons}</View>
-        </View>
-    )
-}
 
 // Styles
 const styles = StyleSheet.create({
@@ -141,10 +43,28 @@ const styles = StyleSheet.create({
     },
 })
 
-// Redux Mappers
-const mapDispatchToProps = dispatch => ({
-    setInitialUserData: (user, jwt) => dispatch(SetInitialUserData(user, jwt)),
-})
+// no_neon_one - "Btw I Use Arch!" (02/17/20)
+const Auth = () => {
+    return (
+        <View style={styles.Container}>
+            <Text style={styles.SectionTitle}>Welcome to Grüvee Beta!</Text>
+            <View style={styles.TextContainer}>
+                <Text style={styles.SectionDetail}>
+                    Thanks for taking part in Grüvee Beta for iOS and Android!
+                </Text>
+                <Text style={styles.SectionDetail}>
+                    It's pretty simple, to get started all you need to do is press the "Login With
+                    Spotify" button.
+                </Text>
+                <Text style={styles.SectionDetail}>
+                    This will give you all the fanciness Grüvee has to offer. Please reach out on
+                    the Discord for any feedback and or questions!
+                </Text>
+            </View>
+            <View style={styles.ButtonContainer}>{Buttons}</View>
+        </View>
+    )
+}
 
 // sillyonly "110 to Get this! I DEMAND A DISCOUNT SINCE I AM A LOYAL CUSTOMER" (02/17/20)
-export default connect(null, mapDispatchToProps)(Auth)
+export default memo(Auth)
